Move code editor constants to util
