+++ frontend/src/lib/CodeEditorUtil.js
